requirements.txt
django_rest_framework
face_recognition
opencv-python
numpy
torch
torchvision
pandas
psutil
pyyaml
tqdm
ultralytics
colorama
imutils
easyocr
